+++ CHANGELOG.md
@@ -16,6 +16,9 @@ FlatLaf Change Log
   Platform Module System (JPMS) for application. (issue #289)
 - Native window decorations: Removed superfluous pixel-line at top of screen
   when window is maximized. (issue #296)
+- Window decorations: Fixed random window title bar background in cases were
+  background is not filled by custom window/rootpane components and
+  `TitlePane.unifiedBackground` is `true`.
 - Button and ToggleButton: Do not paint background of disabled (and unselected)
   toolBar buttons. (issue #292; regression since fixing #112)
 - TabbedPane: Fixed NPE when creating/modifying in another thread. (issue #299)